Rename org to org.freepay
